News-MCP/pyproject.toml
[project]
name = "news-mcp"
version = "0.1.0"
description = "MCP server for latest news"
requires-python = ">=3.11"
dependencies = [
  "fastmcp",
  "python-dotenv",
  "aiohttp"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["news_mcp"]
